poster and trailer moved to movie/components
